rename errors.rs to error.rs
