+++ packages/crypto/src/lib.rs
@@ -2,4 +2,4 @@
 extern crate lazy_static;
 
 pub mod ecvrf;
-pub mod errors;
+pub mod error;
